Another try at CSS-animated loading dots; fixes #25
For posterity: before, the keyframe animation changed a bullet's color.
The animation was applied to each of three elements, with each element
getting a successively-larger animation-delay. The trouble is that the
browser doesn't know that these three (separately-applied) animations
are related, so they can be timed separately from one another.

Solution: add a fourth dot, in the bronze color. The animation applies
only to it, and adjusts its position to cover up each of the other three
dots in turn. Seems to work!